update products in cart, show total quantity and total price in cart

+++ src/app/home/landing/landing.component.ts
@@ -18,7 +18,7 @@ export class LandingComponent implements OnInit {
    }
 
   async ngOnInit() {
-    this.cart = this.cartService.getCart()
+    this.cart = JSON.parse(sessionStorage.getItem('cart'))
     this.products =  await this.dataStorageService.getProductData()
     let object2 = sessionStorage.getItem('cart')
   }
@@ -36,9 +36,20 @@ export class LandingComponent implements OnInit {
   }
 
   addToCart(key, quantity) {
-    let newProduct = this.products.find(x => x.id == key)
-    newProduct.quantity = quantity
-    this.cart.push(newProduct) // adds product object to cart
-    sessionStorage.setItem('cart', JSON.stringify(this.cart))
+    if (this.cart.find(item => item.id == key)) { //if there is already an item in the cart of this product, adds quantity to the item
+      let oldIndex = this.cart.findIndex(item => item.id == key)
+      console.log(this.cart[oldIndex])
+      this.cart[oldIndex].quantity =  this.cart[oldIndex].quantity + parseInt(quantity)
+      this.cart[oldIndex].thisItemTotal = this.cart[oldIndex].quantity * this.cart[oldIndex].productPrice
+      console.log(this.cart[oldIndex])
+      sessionStorage.setItem('cart', JSON.stringify(this.cart))
+    } else {
+      let newProduct = this.products.find(x => x.id == key) //adds new product to cart
+      newProduct.quantity = parseInt(quantity)
+      newProduct.thisItemTotal =  newProduct.quantity * newProduct.productPrice  
+      this.cart.push(newProduct) // adds product object to cart
+      sessionStorage.setItem('cart', JSON.stringify(this.cart))
+    }
+
   }
 }
